Extract `FeedImage` mapping to its own helper class
